Basic Odospinner commit
The Odospinner works, but not well. You can click it and then press the up and down arrow keys to change the value. This also moves the camera even though it's not supposed to. You're also supposed to be able to drag up and down to change the value, but you can't.